requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7
gunicorn==21.2.0
